+++ system/src/sheets/DishonoredBaseActorSheet.mjs
@@ -1,4 +1,5 @@
-export default class DishonoredBaseActorSheet extends ActorSheet {
+export default class DishonoredBaseActorSheet
+	extends foundry.appv1.sheets.ActorSheet {
 
 	/** @override */
 	static get defaultOptions() {
@@ -164,7 +165,8 @@ export default class DishonoredBaseActorSheet extends ActorSheet {
 		context.isNPC = this.actor.type === "npc";
 		context.system = foundry.utils.duplicate(context.data.system);
 
-		context.notesHTML = await TextEditor.enrichHTML(
+		const textEditor = foundry.applications.ux.TextEditor.implementation;
+		context.notesHTML = await textEditor.enrichHTML(
 			this.actor.system.notes,
 			{
 				async: true,